requirements.txt
-i https://pypi.org/simple
exceptiongroup==1.2.0 ; python_version < '3.11'
iniconfig==2.0.0 ; python_version >= '3.7'
mpmath==1.3.0
packaging==23.2 ; python_version >= '3.7'
pluggy==1.4.0 ; python_version >= '3.8'
pytest==8.0.0
tomli==2.0.1 ; python_version < '3.11'
-i https://pypi.org/simple
contourpy==1.2.0 ; python_version >= '3.9'
cycler==0.12.1 ; python_version >= '3.8'
exceptiongroup==1.2.0 ; python_version < '3.11'
environs[django]==9.5.0
whitenoise==6.1.0
gunicorn==20.1.0
djangorestframework==3.14.0
fonttools==4.48.1 ; python_version >= '3.8'
importlib-resources==6.1.1 ; python_version < '3.10'
iniconfig==2.0.0 ; python_version >= '3.7'
kiwisolver==1.4.5 ; python_version >= '3.7'
matplotlib==3.8.2
mpmath==1.3.0
numpy==1.26.4
packaging==23.2 ; python_version >= '3.7'
pillow==10.2.0 ; python_version >= '3.8'
pluggy==1.4.0 ; python_version >= '3.8'
psycopg2-binary==2.9.3
pyparsing==3.1.1 ; python_full_version >= '3.6.8'
pytest==8.0.0
pytest-django==4.5.2
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1 ; python_version < '3.11'
zipp==3.17.0 ; python_version < '3.10'
